% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calccomposite.R
\name{calccomposite}
\alias{calccomposite}
\title{Title Calculate composite score using one of many functions}
\usage{
calccomposite(inputfile)
}
\arguments{
\item{inputfile}{JSON file that contains compVar, which is the name of the composite variable, and args, which contains the named arguments and values that go into the calculations}
}
\value{
JSON formatted compositescore
}
\description{
Title Calculate composite score using one of many functions
}
\examples{
calccomposite( "exampleJSON.json")
}

