% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calccomposite.R
\name{calccomposite_func}
\alias{calccomposite_func}
\title{Title Calculate composite score using lists of functions}
\usage{
calccomposite_func(testname, ...)
}
\arguments{
\item{testname}{Name of the composite score}

\item{...}{Arguments are substituent test scores that are needed in the calculation of the test score. May be named or unnamed (if named, names have to be correct)}
}
\value{
Value of the composite score
}
\description{
Title Calculate composite score using lists of functions
}
\examples{
calccomposite_func( "AVLT__total_1_to_5", AVLT_trial1 = 5, AVLT_trial2 = 5, AVLT_trial3 = 7, AVLT_trial4 = 8 , AVLT_trial5 = 12)
}

